+++ Source/Scripts/Argus/AntorusTheBurningThrone/GarothiWorldbreaker.cs
@@ -229,7 +229,7 @@ namespace Scripts.Argus.AntorusTheBurningThrone.GarothiWorldbreaker
             }
         }
 
-        public override void DamageTaken(Unit attacker, ref uint damage, DamageEffectType damageType, SpellInfo spellInfo = null)
+        public override void DamageTaken(Unit attacker, ref float damage, DamageEffectType damageType, SpellInfo spellInfo = null)
         {
             if (me.HealthBelowPctDamaged(_apocalypseDriveHealthLimit[_apocalypseDriveCount], damage))
             {
